remove v0.1 codebase
This codebase was archived and abandoned.

We'll restart the project with v0.2, just exposing the table of
multicodecs as Go constants. The first step needed for that is to remove
all existing code.

It's useful to keep the same git history and tags, so that we can reuse
the same repository without breaking any users of previous versions.
It's also nice to keep the historical data in a single place.