replace Of with Code.Set, implementing flag.Value
This means one can use a multicodec.Code directly with flag.Var, which
will do the right thing when parsing and printing multicodec names.

The implementation of Set is the same as Of, with the only difference
that it's a method modifying the receiver, and returns an error instead
of a boolean.

While at it, rearrange the code a bit. We were sticking extra static
code in gen.go just to have it in code.go, while also having a separate
doc.go.

Now, gen.go simply generates the table constants in code_table.go, and
code.go includes the type, its non-generated methods, and the package
docs.